requirements.txt
Babel==1.3
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.6
coverage==3.7.1
docutils==0.12
mock==1.0.1
nose==1.3.6
pluggy==0.3.0
py==1.4.30
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
requests-toolbelt==0.4.0
responses==0.4.0
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.8
tox==2.0.1
virtualenv==13.1.0
